+++ test/src/integrations/http_tracking_client_test.dart
@@ -29,8 +29,8 @@ class _RecordingRouter implements TelemetryRouter {
   @override
   void ingest(
     TelemetryItem item, {
+    required SessionActivityKind activity,
     bool skipBuffer = false,
-    SessionActivityKind activity = SessionActivityKind.active,
   }) {
     ingested.add(item);
   }